requirements.txt
selenium
psycopg2-binary
python-dotenv
pandas